+++ spec/unit/mongoid/matchers_spec.rb
@@ -105,7 +105,11 @@ describe Mongoid::Matchers do
     context "when performing complex matching" do
 
       let(:document) do
-        Address.new(:services => ["first", "second"], :number => 100)
+        Address.new(
+          :services => ["first", "second"],
+          :number => 100,
+          :map => { :key => "value" }
+        )
       end
 
       context "with an $all selector" do
@@ -425,6 +429,31 @@ describe Mongoid::Matchers do
 
           let(:selector) do
             { :services => { "$size" => 5 } }
+          end
+
+          it "returns false" do
+            document.matches?(selector).should be_false
+          end
+        end
+      end
+
+      context "with a hash value" do
+
+        context "when the attributes match" do
+
+          let(:selector) do
+            { :map => { :key => "value" } }
+          end
+
+          it "returns true" do
+            document.matches?(selector).should be_true
+          end
+        end
+
+        context "when the attributes do not match" do
+
+          let(:selector) do
+            { :map => { :key => "value2" } }
           end
 
           it "returns false" do